Adjusted VertexBuffer to accept QuadVertex with mix of float and int variables

+++ mmd-editor/src/main/java/dev/xfj/EditorLayer.java
@@ -177,7 +177,7 @@ public class EditorLayer extends Layer {
         int mouseY = (int) mousePosition.y;
 
         if (mouseX >= 0 && mouseY >= 0 && mouseX < (int) viewportSize.x && mouseY < (int) viewportSize.y) {
-            float pixelData = framebuffer.readPixel(1, mouseX, mouseY);
+            int pixelData = framebuffer.readPixel(1, mouseX, mouseY);
             hoveredEntity = activeScene.getEntityById(pixelData);
         }
